removing type: ignore
